% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{admixplorer}
\alias{admixplorer}
\title{Main wrapper function for admixplorer analysis}
\usage{
admixplorer(
  infile,
  outfile,
  method = "GLOBETROTTER",
  ks = "1,2,3,4",
  mcmc_chains = 3,
  sample_age_est = TRUE,
  plot = TRUE,
  apply_date_filter = TRUE
)
}
\arguments{
\item{infile}{Path to input file}

\item{outfile}{Output file prefix}

\item{method}{Method to use ("GLOBETROTTER" or "DATES")}

\item{ks}{Comma-separated string of k values to try (default "1,2,3,4")}

\item{mcmc_chains}{Number of MCMC chains per k (default 3)}

\item{sample_age_est}{Whether to estimate sample ages (default TRUE)}

\item{plot}{Whether to create plots (default TRUE)}
}
\value{
Combined output dataframe
}
\description{
Main wrapper function for admixplorer analysis
}
