% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{calculate_final_likelihoods}
\alias{calculate_final_likelihoods}
\title{Calculate final likelihoods with different sampling age approaches}
\usage{
calculate_final_likelihoods(
  cluster_assignments,
  final_cluster_means,
  best_cluster_means_all,
  std_errors,
  dates,
  age_summaries,
  best_sampling_ages,
  n
)
}
\description{
Calculate final likelihoods with different sampling age approaches
}
\keyword{internal}
