% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{calculate_fst_scaling}
\alias{calculate_fst_scaling}
\title{Calculate FST-based scaling factor}
\usage{
calculate_fst_scaling(input_fst, method)
}
\arguments{
\item{input_fst}{FST between admixing populations}

\item{method}{Method used (GLOBETROTTER or DATES)}
}
\value{
List with scaling information
}
\description{
Calculate FST-based scaling factor
}
\keyword{internal}
