% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{calculate_joint_date_stats}
\alias{calculate_joint_date_stats}
\title{Calculate joint date statistics for individuals}
\usage{
calculate_joint_date_stats(n, cluster_assignments, cluster_means, cluster_sds)
}
\description{
Calculate joint date statistics for individuals
}
\keyword{internal}
