% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{calculate_likelihood_improvements}
\alias{calculate_likelihood_improvements}
\title{Calculate likelihood improvements between k values}
\usage{
calculate_likelihood_improvements(all_mcmc_results, n_individuals)
}
\arguments{
\item{all_mcmc_results}{List of MCMC results from run_clustering_analysis}

\item{n_individuals}{Number of individuals (for per-individual likelihood)}
}
\value{
List of improvements and related statistics
}
\description{
Calculate likelihood improvements between k values
}
