% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{calculate_uncertainties}
\alias{calculate_uncertainties}
\title{Calculate uncertainties for sampling ages and joint dates}
\usage{
calculate_uncertainties(
  n,
  k,
  cluster_assignments,
  sampling_age_samples,
  cluster_mean_samples,
  final_cluster_sds,
  sample_ages
)
}
\description{
Calculate uncertainties for sampling ages and joint dates
}
\keyword{internal}
