% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{check_install_packages}
\alias{check_install_packages}
\title{Check and install required packages}
\usage{
check_install_packages(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\description{
Check and install required packages
}
\keyword{internal}
