% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{clustering_strength}
\alias{clustering_strength}
\title{Calculate clustering strength from co-clustering matrix}
\usage{
clustering_strength(coincidence_matrix)
}
\arguments{
\item{coincidence_matrix}{Co-clustering matrix from MCMC results}
}
\value{
Clustering strength score between 0 and 1
}
\description{
Calculate clustering strength from co-clustering matrix
}
