% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{create_result_dataframe}
\alias{create_result_dataframe}
\title{Create final result dataframe}
\usage{
create_result_dataframe(
  n,
  cluster_assignments,
  age_summaries,
  best_sampling_ages,
  joint_date_stats,
  joint_date_stats_best,
  uncertainty_results,
  ll_results,
  acceptance_rate,
  mean_log_likelihood_lambda1,
  mean_log_likelihood_lambda10,
  combined_stats,
  combined_stats_best
)
}
\description{
Create final result dataframe
}
\keyword{internal}
