% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_final_output}
\alias{generate_final_output}
\title{Generate final output}
\usage{
generate_final_output(
  all_mcmc_results,
  recommended_k,
  original_data,
  pop.vec,
  dates_original,
  sample_age_est,
  outfile,
  method,
  fst,
  thresholds_info
)
}
\arguments{
\item{all_mcmc_results}{List of MCMC results}

\item{recommended_k}{Recommended k value}

\item{original_data}{Original filtered data}

\item{pop.vec}{Population vector}

\item{dates_original}{Original dates}

\item{sample_age_est}{Whether sample ages were estimated}

\item{outfile}{Output file prefix}

\item{method}{Method used}

\item{fst}{FST value used}

\item{thresholds_info}{Threshold information from analysis}
}
\value{
Combined output dataframe
}
\description{
Generate final output
}
