% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_clustering.R
\name{mcmc_clustering}
\alias{mcmc_clustering}
\title{MCMC clustering for admixture dates}
\usage{
mcmc_clustering(
  dates,
  std_errors,
  init_cluster_means,
  cluster_assignments,
  num_iter = 2e+05,
  burn_in = 1e+05,
  k = 2,
  thin = 100,
  sample_ages = FALSE,
  age_ranges = NULL,
  sampling_ages = NULL,
  lambda = 10
)
}
\arguments{
\item{dates}{Vector of admixture dates}

\item{std_errors}{Standard errors for dates}

\item{init_cluster_means}{Initial cluster means}

\item{cluster_assignments}{Initial cluster assignments}

\item{num_iter}{Number of MCMC iterations}

\item{burn_in}{Number of burn-in iterations}

\item{k}{Number of clusters}

\item{thin}{Thinning interval}

\item{sample_ages}{Whether to sample ages}

\item{age_ranges}{Age range matrix}

\item{sampling_ages}{Initial sampling ages}

\item{lambda}{Lambda parameter}
}
\value{
List with MCMC results
}
\description{
MCMC clustering for admixture dates
}
\keyword{internal}
