% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_clustering_data}
\alias{prepare_clustering_data}
\title{Prepare clustering data}
\usage{
prepare_clustering_data(data, sample_age_est = TRUE)
}
\arguments{
\item{data}{Filtered dataframe from read_and_filter_data}

\item{sample_age_est}{Whether to estimate sample age or use mean}
}
\value{
List of prepared data for clustering
}
\description{
Prepare clustering data
}
