% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{print_time_traveler_summary}
\alias{print_time_traveler_summary}
\title{Print time traveler summary}
\usage{
print_time_traveler_summary(time_traveler_log, k)
}
\description{
Print time traveler summary
}
\keyword{internal}
