% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{process_mcmc_results}
\alias{process_mcmc_results}
\title{Process MCMC results after sampling (MAIN PROCESSING FUNCTION)}
\usage{
process_mcmc_results(
  n,
  k,
  dates,
  std_errors,
  best_cluster_assignments,
  best_sampling_ages,
  best_cluster_means,
  sampling_age_samples,
  cluster_mean_samples,
  co_clustering_matrix,
  acceptance_rate,
  num_iter,
  burn_in,
  thin,
  sample_ages,
  age_ranges,
  time_traveler_log,
  mean_log_likelihood_lambda1,
  mean_log_likelihood_lambda10
)
}
\description{
Process MCMC results after sampling (MAIN PROCESSING FUNCTION)
}
\keyword{internal}
