% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clustering.R
\name{process_single_k}
\alias{process_single_k}
\title{Process single k value clustering}
\usage{
process_single_k(
  k_current,
  prepared_data,
  chains,
  sample_age_est,
  plot,
  outfile
)
}
\arguments{
\item{k_current}{Current k value to process}

\item{prepared_data}{List from prepare_clustering_data}

\item{chains}{Number of MCMC chains}

\item{sample_age_est}{Whether to estimate sample ages}

\item{plot}{Whether to create plots}

\item{outfile}{Output file prefix}
}
\value{
MCMC results for this k value
}
\description{
Process single k value clustering
}
\keyword{internal}
