% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{run_clustering_analysis}
\alias{run_clustering_analysis}
\title{Run clustering analysis for all k values}
\usage{
run_clustering_analysis(
  prepared_data,
  ks,
  chains,
  sample_age_est,
  plot,
  outfile
)
}
\arguments{
\item{prepared_data}{List from prepare_clustering_data}

\item{ks}{Vector of k values to try}

\item{chains}{Number of MCMC chains}

\item{sample_age_est}{Whether to estimate sample ages}

\item{plot}{Whether to create plots}

\item{outfile}{Output file prefix}
}
\value{
List of MCMC results for each k
}
\description{
Run clustering analysis for all k values
}
