% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_and_filter_data}
\alias{read_and_filter_data}
\title{Read and filter admixture data}
\usage{
read_and_filter_data(filename, apply_date_filter = TRUE)
}
\arguments{
\item{filename}{Path to input file}

\item{apply_date_filter}{Logical, whether to apply date-based outlier filters (default TRUE)}
}
\value{
List containing filtered dataframe and original individuals

List containing filtered dataframe and original individuals
}
\description{
Read and filter admixture data
}
