% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thresholds_config}
\alias{thresholds_config}
\title{Threshold configuration data}
\format{
A list with 2 elements:
\describe{
\item{GLOBETROTTER}{List of thresholds for GLOBETROTTER method}
\item{DATES}{List of thresholds for DATES method}
}
}
\usage{
thresholds_config
}
\description{
A list containing threshold values for different methods (GLOBETROTTER and DATES)
used in likelihood-based k selection.
}
\keyword{datasets}
